+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This is a changelog for **ZachranObed** application.
 - **ZOB-368** Reorganize image assets and remove unused code.
 - **ZOB-371** Add custom lint rules to support architecture.
 - **ZOB-378** Move DTOs and services.
+- **ZOB-379** Move common UI widgets.
 
 ### Removed